Fixed reporting consent issue
Previously, if the `~/.activeloop` directory did not exist, the
reporting consent configuration file was not being saved and so consent
checks were all coming through as False.

Fixed this problem by creating the directory containing the consent file
if it does not exist. This mirrors how tokens are stored when a user
signs in with `activeloop login`.