Adds "machine_id:<original client id>" tag to reports
I also reverted the change where client_id is overwritten by username
when the config file is written, as we need to retain a distinct
machine_id for each machine the user uses `activeloop` on.

This remapping of keys inside reporting_config should probably be
refactored at some point because it could be confusing for future
builders on this project. For now, it was the simplest way to get the
desired behavior in the analytics data. The more significant (and
correct) change will require more testing than this.