Do not hide S3 access errors
If e.g. the wrong AWS profile is used the download of the open datasets
which are stored in S3 will fail. These errors was completely hidden
and was instead displayed as if the dataset does not exists.

This commit creates a new excpetion type which is not excepted to
make it clear that it is an AWS S3 access error that is the cause.