Variable local cache prefix (#1839)
* Allows for changing the LOCAL_CACHE_PREFIX via an environment variable.

Thus allowing for greater flexibility when using the local cache in the PyTorch data loader. Such as using a mounted filesystem such as /scratch or another location, i.e. /tmp, instead of the home directory.

* Updated PyTorch function documentation

Changed to reflect the ability to use LOCAL_CACHE_PREFIX as an environment variable.

* Update hub/util/cache_chain.py

Simplify the assignment of local_cache_prefix not to require an if then else statement. Thus making the code easier to read.

Co-authored-by: David Isayan <d.isayan@gmail.com>

* Update hub/util/storage.py

Simplify the assignment of local_cache_prefix not to require an if then else statement. Thus making the code easier to read.

Co-authored-by: David Isayan <d.isayan@gmail.com>

* Removal of if then else statement for local_cache_prefix

This is no longer required as the os.getenv default value is used instead.

Co-authored-by: David Isayan <d.isayan@gmail.com>