Do not hide S3 access errors (#1884)
* fix

* Do not hide S3 access errors

If e.g. the wrong AWS profile is used the download of the open datasets
which are stored in S3 will fail. These errors was completely hidden
and was instead displayed as if the dataset does not exists.

This commit creates a new excpetion type which is not excepted to
make it clear that it is an AWS S3 access error that is the cause.

* Handle repeated samples in shuffle (#2018)

* fix

* Fix.

* fix for fix

Co-authored-by: Sasun Hambardzumyan <xustup@gmail.com>

Co-authored-by: FayazRahman <fayazrahman4u@gmail.com>
Co-authored-by: Abhinav Tuli <42538472+AbhinavTuli@users.noreply.github.com>
Co-authored-by: Sasun Hambardzumyan <xustup@gmail.com>
Co-authored-by: Fariz Rahman <farizrahman4u@gmail.com>